requirements.txt
adrf==0.1.8
asgiref==3.8.1
async-property==0.2.2
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.12.14
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.0
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
Django==5.1.4
django-channels==0.7.0
django-configurations==2.5.1
django-cors-headers==4.6.0
django-filter==25.1
djangorestframework==3.15.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.0
oauthlib==3.2.2
pillow==11.1.0
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyOpenSSL==24.3.0
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
service-identity==24.2.0
setuptools==75.6.0
six==1.17.0
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.3.0
websockets==14.1
zope.interface==7.2
